Move params to API model
